lock when writing the file on the disk

+++ Countries-API/Helpers/CountriesHelper.cs
@@ -12,6 +12,7 @@ namespace Countries_API.Helpers
 {
     public class CountriesHelper
     {
+        private static readonly object locker = new object();
         public static async Task<byte[]> DownloadImageAsync(string directoryPath, string fileName, Uri uri)
         {
             using var httpClient = new HttpClient();
@@ -33,21 +34,35 @@ namespace Countries_API.Helpers
             return imageBytes;
         }
 
-        public static byte[] DownloadImageAndSaveOnDisk(string directoryPath, string fileName, string fileUrl, string fileExtension)
+        public static byte[] DownloadImageAndSaveOnDisk(string directoryPath, string countryIsoCode, string fileUrl, string fileExtension)
         {
             using(WebClient webClient = new WebClient())
             {
-                // Download the image and write to the file
-                var imageBytes = webClient.DownloadData(fileUrl);
+                //Only a single request should be allowed to write the file to the disk
+                //lock in order to have concurrency control
+                lock (locker)
+                {
+                    #region Concurrency Control, Check again if file is on the disk
+                    //In case two same requests occur at the same time, check again if the file is already written on disk
+                    if (File.Exists("ImageData\\" + countryIsoCode + fileExtension))
+                    {
+                        var fileOnDisk = File.ReadAllBytes("ImageData\\" + countryIsoCode + fileExtension);
+                        return fileOnDisk;
+                    }
+                    #endregion
 
-                // Create file path and ensure directory exists
-                var path = Path.Combine(directoryPath, $"{fileName}{fileExtension}");
-                Directory.CreateDirectory(directoryPath);
+                    // Download the image and write to the file
+                    var imageBytes = webClient.DownloadData(fileUrl);
 
-                //Write the file on disk
-                File.WriteAllBytes(path, imageBytes);
+                    // Create file path and ensure directory exists
+                    var path = Path.Combine(directoryPath, $"{countryIsoCode}{fileExtension}");
+                    Directory.CreateDirectory(directoryPath);
 
-                return imageBytes;
+                    //Write the file on disk
+                    File.WriteAllBytes(path, imageBytes);
+
+                    return imageBytes;
+                }                                
             }
         }
